+++ src/wallets/index.ts
@@ -26,8 +26,12 @@ import {
   GetWalletResponse,
   GetWalletDepositInstructionsRequest,
   GetWalletDepositInstructionsResponse,
+  ListWalletAddressesRequest,
+  ListWalletAddressesResponse,
   CreateWalletRequest,
   CreateWalletResponse,
+  CreateWalletDepositAddressRequest,
+  CreateWalletDepositAddressResponse,
 } from './types';
 
 export interface IWalletsService {
@@ -54,11 +58,29 @@ export interface IWalletsService {
     | CoinbasePrimeException
   >;
 
+  listWalletAddresses(
+    request: ListWalletAddressesRequest,
+    options?: CoinbaseCallOptions
+  ): Promise<
+    | ListWalletAddressesResponse
+    | CoinbasePrimeClientException
+    | CoinbasePrimeException
+  >;
+
   createWallet(
     request: CreateWalletRequest,
     options?: CoinbaseCallOptions
   ): Promise<
     CreateWalletResponse | CoinbasePrimeClientException | CoinbasePrimeException
+  >;
+
+  createWalletDepositAddress(
+    request: CreateWalletDepositAddressRequest,
+    options?: CoinbaseCallOptions
+  ): Promise<
+    | CreateWalletDepositAddressResponse
+    | CoinbasePrimeClientException
+    | CoinbasePrimeException
   >;
 }
 
@@ -113,6 +135,36 @@ export class WalletsService implements IWalletsService {
     return response.data as GetWalletDepositInstructionsResponse;
   }
 
+  async listWalletAddresses(
+    request: ListWalletAddressesRequest,
+    options?: CoinbaseCallOptions
+  ): Promise<
+    | ListWalletAddressesResponse
+    | CoinbasePrimeClientException
+    | CoinbasePrimeException
+  > {
+    let queryParams: Record<string, string | number> = {};
+    if (request.networkId) {
+      queryParams['network.id'] = request.networkId;
+    }
+    if (request.networkType) {
+      queryParams['network.type'] = request.networkType;
+    }
+    if (request.cursor) {
+      queryParams.cursor = request.cursor;
+    }
+    if (request.limit) {
+      queryParams.limit = request.limit;
+    }
+    const response = await this.client.request({
+      url: `portfolios/${request.portfolioId}/wallets/${request.walletId}/addresses`,
+      queryParams,
+      callOptions: options,
+    });
+
+    return response.data as ListWalletAddressesResponse;
+  }
+
   async createWallet(
     request: CreateWalletRequest,
     options?: CoinbaseCallOptions
@@ -128,5 +180,30 @@ export class WalletsService implements IWalletsService {
     });
 
     return response.data as CreateWalletResponse;
+  }
+
+  async createWalletDepositAddress(
+    request: CreateWalletDepositAddressRequest,
+    options?: CoinbaseCallOptions
+  ): Promise<
+    | CreateWalletDepositAddressResponse
+    | CoinbasePrimeClientException
+    | CoinbasePrimeException
+  > {
+    let queryParams: Record<string, string | number> = {};
+    if (request.networkId) {
+      queryParams['network.id'] = request.networkId;
+    }
+    if (request.networkType) {
+      queryParams['network.type'] = request.networkType;
+    }
+    const response = await this.client.request({
+      url: `portfolios/${request.portfolioId}/wallets/${request.walletId}/addresses`,
+      queryParams,
+      method: Method.POST,
+      callOptions: options,
+    });
+
+    return response.data as CreateWalletDepositAddressResponse;
   }
 }